chore: move layout components from `theme/layouts` to `components/templates` folder
