+++ components/templates/basic.tsx
@@ -4,30 +4,31 @@ import {useSafeArea} from 'react-native-safe-area-context';
 
 import {SPACING_TOP, SPACING_BOTTOM, SPACING_HORIZONTAL} from './shared';
 
-import {Heading} from '../../components/atoms/heading';
+import {Heading} from '../atoms/heading';
 import {colors} from '../../constants/colors';
 
 interface LayoutProps {
   heading?: string;
-  children: any;
+  backgroundColor?: string;
+  children: React.ReactNode;
 }
 
-export const PinnedBottom: FC<LayoutProps> = ({children, heading}) => {
+export const Basic: FC<LayoutProps> = ({
+  children,
+  heading,
+  backgroundColor
+}) => {
   const insets = useSafeArea();
-  const content = React.Children.toArray(children);
-  const bottom = content.pop();
 
   return (
     <View
       style={[
         styles.container,
-        {paddingBottom: insets.bottom + SPACING_BOTTOM}
+        {paddingBottom: insets.bottom + SPACING_BOTTOM},
+        !!backgroundColor && {backgroundColor}
       ]}>
-      <View>
-        {heading && <Heading accessibilityFocus text={heading} />}
-        {content}
-      </View>
-      <View>{bottom}</View>
+      {heading && <Heading accessibilityFocus text={heading} />}
+      {children}
     </View>
   );
 };
@@ -36,7 +37,6 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: colors.white,
-    justifyContent: 'space-between',
     paddingTop: SPACING_TOP,
     paddingHorizontal: SPACING_HORIZONTAL
   }
