+++ components/templates/shared.ts
