+++ components/views/upload-keys.tsx
@@ -22,7 +22,7 @@ import {Toast} from '../atoms/toast';
 import {CodeInput} from '../molecules/code-input';
 
 import {colors} from '../../constants/colors';
-import Layouts from '../../theme/layouts';
+import Layouts from '../templates';
 import {text, baseStyles} from '../../theme';
 
 type UploadStatus =
